Set up basic backend + react app
A very basic backend has been set up in the backend folder. Axios will be used later in the frontend to send requests to the backend.

The database currently supports two models: Users and Debates
Users have a username field.
Debates have topics, users, and date.
Debates will probably have an array of strings in the future for the text conversation.

I also created a basic default react app located in the debchat folder, and that's where the frontend will be. This folder can be deleted and the backend should not suffer.