Updated Debate schemas + added dependencies to package.json
Updated Debate schemas to better fit needs. Split users into proponent and opponent, and added the conversation field that will store text messages.

The type for conversation is currently an Array of Strings, but this can change depending on how text messages and chatboxes work in React.

After installing dependencies on my local system, the package.json file is updated as appropriate.