Testing Chatbox works with socket.io
I implemented a sample chatbox that uses socket.io to allow multiple users to receive new chats as they are posted to the database. It's listed as a table for now, but will be touched up later into a prettier UI.

Currently, it's hardcoded to a specific Debate, and users are represented by their socket ID instead of their names. Also, all chats sent will be sent to the respective debate, but the server will emit universally to all sockets. This means that a chat in a separate debate will be told to update its chatbox.

Next, rooms will be tested to see if we can remove the universal emits, though this isn't necessarily a problem. Data won't bounce between debates and corrupt other chats.

Dependencies in package.json have been updated.