Integrated Auth0 User Authorization
Integrated Auth0 user authorization

Added a Log In button that redirects to Auth0's default Universal Login page. User credentials are stored on their databases, and verification of authentication is stored as a cookie on the browser.

If, during testing, a HTTP 400 Error appears, clear cookies and the issue should resolve.

Messages sent in the ChatBox now have users properly assigned to each message, given that they have signed in.

Current issues/updates:
- it is still possible for those who have not logged in to access all pages of the website via the URL. If users access forbidden links, they should be redirected to the Auth0 login page, but that fix has not been successful yet
- once a custom domain is secured for debchat, we'll look into embedding the login page instead of redirecting to Auth0's universal login
- User Profile is very minimal